Code and comments
Hi Leo, this is the code of the LondonR workshop I attended, I went through all the lines and added functions and their explanations/comments. It is an overview of string and words manipulation, word frequency, sentiment analysis and web scraping. I think it covers some of the topics of the the https://www.tidytextmining.com/ you shared with me some times ago. I did not not understand three lines (144, 159, 225). Feel free to merge this branch with the main one and then modify/comment with another pull request if you like :) Thanks, Elena